requirements.txt
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
fonttools==4.57.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
nltk==3.9.1
numpy==1.24.4
oauthlib==3.2.2
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
six==1.17.0
textblob==0.18.0.post0
tqdm==4.67.1
tweepy==4.14.0
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
